Refactor tasks module: migrate `app.tasks` to `app.jobs`, update all imports and references, and adjust related tests and documentation accordingly.
